Add dependncy so flex/bison builda first. Refactor code so main is just a stub. This gives more test coverage
