+++ CMakePasm/source/parseargs.h
@@ -1,4 +1,4 @@
 #pragma once
 
+int pre_parse();
 int parse_arguments(int argc, char* argv[]);
-
