+++ CMakePasm/source/pass.h
@@ -1,0 +1,9 @@
+#pragma once
+
+extern void reset_lex(void);
+extern void parse_pass(void);
+extern int assemble(void);
+
+
+
+
